Check for 22.04 queue.
