+++ src/tools/ScenarioMeasurement/Startup/PerfLabValues.cs
@@ -1,7 +1,9 @@
-﻿namespace ScenarioMeasurement;
+﻿using System;
+
+namespace ScenarioMeasurement;
 
 // see src\scenarios\staticdeps\PerfLab.cs
-class PerfLabValues
+static class PerfLabValues
 {
     // to "prevent" event id collisions
     private const int MagicConstant = 6666;
@@ -14,4 +16,18 @@ class PerfLabValues
     public const string ForwarderName = "PerfLabGenericEventSourceForwarder";
     public const string LTTngProviderName = "PerfLabGenericEventSourceLTTngProvider";
     public const string LTTngProviderLibraryName = LTTngProviderName;
+
+    public static class SharedHelpers
+    {
+        public static bool IsUbuntu22Queue()
+        {
+            var queue = Environment.GetEnvironmentVariable("PERFLAB_QUEUE");
+            if (string.IsNullOrWhiteSpace(queue))
+            {
+                return false;
+            }
+            return queue.Contains("ubuntu", StringComparison.OrdinalIgnoreCase)
+                && queue.Contains("22", StringComparison.Ordinal);
+        }
+    }
 }
